merge upstream

+++ controllers/grid/users/chapter/ChapterGridHandler.inc.php
@@ -3,9 +3,9 @@
 /**
  * @file controllers/grid/users/chapter/ChapterGridHandler.inc.php
  *
- * Copyright (c) 2014-2016 Simon Fraser University Library
- * Copyright (c) 2000-2016 John Willinsky
- * Distributed under the GNU GPL v2. For full terms see the file docs/COPYING.
+ * Copyright (c) 2014-2020 Simon Fraser University
+ * Copyright (c) 2000-2020 John Willinsky
+ * Distributed under the GNU GPL v3. For full terms see the file docs/COPYING.
  *
  * @class ChapterGridHandler
  * @ingroup controllers_grid_users_chapter
@@ -31,15 +31,19 @@ class ChapterGridHandler extends CategoryGridHandler {
 	/**
 	 * Constructor
 	 */
-	function ChapterGridHandler() {
-		parent::CategoryGridHandler();
+	function __construct() {
+		parent::__construct();
 		$this->addRoleAssignment(
 			array(ROLE_ID_AUTHOR, ROLE_ID_SUB_EDITOR, ROLE_ID_MANAGER, ROLE_ID_ASSISTANT),
 			array(
 				'fetchGrid', 'fetchRow', 'fetchCategory', 'saveSequence',
-				'addChapter', 'editChapter', 'updateChapter', 'deleteChapter',
+				'addChapter', 'editChapter', 'editChapterTab', 'updateChapter', 'deleteChapter',
 				'addAuthor', 'editAuthor', 'updateAuthor', 'deleteAuthor'
 			)
+		);
+		$this->addRoleAssignment(
+			array(ROLE_ID_SUB_EDITOR, ROLE_ID_MANAGER, ROLE_ID_ASSISTANT),
+			array('identifiers', 'updateIdentifiers', 'clearPubId',)
 		);
 		$this->addRoleAssignment(ROLE_ID_REVIEWER, array('fetchGrid', 'fetchRow'));
 	}
@@ -54,6 +58,14 @@ class ChapterGridHandler extends CategoryGridHandler {
 	 */
 	function getMonograph() {
 		return $this->getAuthorizedContextObject(ASSOC_TYPE_MONOGRAPH);
+	}
+
+	/**
+	 * Get the publication associated with this chapter grid.
+	 * @return Publication
+	 */
+	function getPublication() {
+		return $this->getAuthorizedContextObject(ASSOC_TYPE_PUBLICATION);
 	}
 
 	/**
@@ -83,21 +95,24 @@ class ChapterGridHandler extends CategoryGridHandler {
 	 * @param $roleAssignments array
 	 */
 	function authorize($request, &$args, $roleAssignments) {
-		import('lib.pkp.classes.security.authorization.SubmissionAccessPolicy');
-		$this->addPolicy(new SubmissionAccessPolicy($request, $args, $roleAssignments));
+		import('lib.pkp.classes.security.authorization.PublicationAccessPolicy');
+		$this->addPolicy(new PublicationAccessPolicy($request, $args, $roleAssignments));
 		return parent::authorize($request, $args, $roleAssignments);
 	}
 
 	/**
-	 * Configure the grid
-	 * @param $request PKPRequest
+	 * @copydoc CategoryGridHandler::initialize()
 	 */
-	function initialize($request) {
-		parent::initialize($request);
+	function initialize($request, $args = null) {
+		parent::initialize($request, $args);
 
 		$this->setTitle('submission.chapters');
 
-		AppLocale::requireComponents(LOCALE_COMPONENT_APP_DEFAULT, LOCALE_COMPONENT_PKP_DEFAULT, LOCALE_COMPONENT_APP_SUBMISSION);
+		AppLocale::requireComponents(LOCALE_COMPONENT_APP_DEFAULT, LOCALE_COMPONENT_PKP_DEFAULT, LOCALE_COMPONENT_APP_SUBMISSION, LOCALE_COMPONENT_PKP_SUBMISSION);
+
+		if ($this->getPublication()->getData('status') === STATUS_PUBLISHED) {
+			$this->setReadOnly(true);
+		}
 
 		if (!$this->getReadOnly()) {
 			// Grid actions
@@ -120,7 +135,7 @@ class ChapterGridHandler extends CategoryGridHandler {
 
 		// Columns
 		// reuse the cell providers for the AuthorGrid
-		$cellProvider = new PKPAuthorGridCellProvider();
+		$cellProvider = new PKPAuthorGridCellProvider($this->getPublication());
 		$this->addColumn(
 			new GridColumn(
 				'name',
@@ -155,13 +170,10 @@ class ChapterGridHandler extends CategoryGridHandler {
 	 * @see GridHandler::initFeatures()
 	 */
 	function initFeatures($request, $args) {
-		$monograph = $this->getMonograph();
-		$userRoles = $this->getAuthorizedContextObject(ASSOC_TYPE_USER_ROLES);
-
-		if ($monograph->getDateSubmitted() == null || array_intersect(array(ROLE_ID_MANAGER, ROLE_ID_SUB_EDITOR), $userRoles)) {
+		if ($this->canAdminister($request->getUser())) {
 			$this->setReadOnly(false);
 			import('lib.pkp.classes.controllers.grid.feature.OrderCategoryGridItemsFeature');
-			return array(new OrderCategoryGridItemsFeature(ORDER_CATEGORY_GRID_CATEGORIES_AND_ROWS));
+			return array(new OrderCategoryGridItemsFeature(ORDER_CATEGORY_GRID_CATEGORIES_AND_ROWS, true, $this));
 		} else {
 			$this->setReadOnly(true);
 			return array();
@@ -172,11 +184,43 @@ class ChapterGridHandler extends CategoryGridHandler {
 	 * @see GridDataProvider::getRequestArgs()
 	 */
 	function getRequestArgs() {
-		$monograph = $this->getMonograph();
 		return array_merge(
 			parent::getRequestArgs(),
-			array('submissionId' => $monograph->getId())
+			array(
+				'submissionId' => $this->getMonograph()->getId(),
+				'publicationId' => $this->getPublication()->getId(),
+			)
 		);
+	}
+
+	/**
+	 * Determines if there should be add/edit actions on this grid.
+	 * @param $user User
+	 * @return boolean
+	 */
+	function canAdminister($user) {
+		$submission = $this->getMonograph();
+		$publication = $this->getPublication();
+		$userRoles = $this->getAuthorizedContextObject(ASSOC_TYPE_USER_ROLES);
+
+		if ($publication->getData('status') === STATUS_PUBLISHED) {
+			return false;
+		}
+
+		if (in_array(ROLE_ID_SITE_ADMIN, $userRoles)) {
+			return true;
+		}
+
+		// Incomplete submissions can be edited. (Presumably author.)
+		if ($submission->getDateSubmitted() == null) return true;
+
+		// The user may not be allowed to edit the metadata
+		if (Services::get('submission')->canEditPublication($submission->getId(), $user->getId())) {
+			return true;
+		}
+
+		// Default: Read-only.
+		return false;
 	}
 
 	/**
@@ -191,10 +235,9 @@ class ChapterGridHandler extends CategoryGridHandler {
 	 * @see GridHandler::loadData
 	 */
 	function loadData($request, $filter) {
-		$monograph = $this->getMonograph();
-		$chapterDao = DAORegistry::getDAO('ChapterDAO');
-		$chapters = $chapterDao->getChapters($monograph->getId());
-		return $chapters->toAssociativeArray();
+		return DAORegistry::getDAO('ChapterDAO')
+			->getByPublicationId($this->getPublication()->getId())
+			->toAssociativeArray();
 	}
 
 
@@ -212,7 +255,9 @@ class ChapterGridHandler extends CategoryGridHandler {
 	 * @see GridHandler::setDataElementSequence()
 	 */
 	function setDataElementSequence($request, $chapterId, $chapter, $newSequence) {
-		$chapterDao = DAORegistry::getDAO('ChapterDAO');
+		if (!$this->canAdminister($request->getUser())) return;
+
+		$chapterDao = DAORegistry::getDAO('ChapterDAO'); /* @var $chapterDao ChapterDAO */
 		$chapter->setSequence($newSequence);
 		$chapterDao->updateObject($chapter);
 	}
@@ -226,7 +271,7 @@ class ChapterGridHandler extends CategoryGridHandler {
 	 */
 	function getCategoryRowInstance() {
 		$monograph = $this->getMonograph();
-		$row = new ChapterGridCategoryRow($monograph, $this->getReadOnly());
+		$row = new ChapterGridCategoryRow($monograph, $this->getPublication(), $this->getReadOnly());
 		import('controllers.grid.users.chapter.ChapterGridCategoryRowCellProvider');
 		$row->setCellProvider(new ChapterGridCategoryRowCellProvider());
 		return $row;
@@ -235,7 +280,7 @@ class ChapterGridHandler extends CategoryGridHandler {
 	/**
 	 * @see CategoryGridHandler::loadCategoryData()
 	 */
-	function loadCategoryData($request, &$chapter, $filter) {
+	function loadCategoryData($request, &$chapter, $filter = null) {
 		$authorFactory = $chapter->getAuthors(); /* @var $authorFactory DAOResultFactory */
 		return $authorFactory->toAssociativeArray();
 	}
@@ -243,23 +288,25 @@ class ChapterGridHandler extends CategoryGridHandler {
 	/**
 	 * @see CategoryGridHandler::getDataElementInCategorySequence()
 	 */
-	function getDataElementInCategorySequence($categoryId, $author) {
+	function getDataElementInCategorySequence($categoryId, &$author) {
 		return $author->getSequence();
 	}
 
 	/**
 	 * @see CategoryGridHandler::setDataElementInCategorySequence()
 	 */
-	function setDataElementInCategorySequence($chapterId, $author, $newSequence) {
+	function setDataElementInCategorySequence($chapterId, &$author, $newSequence) {
+		if (!$this->canAdminister(Application::get()->getRequest()->getUser())) return;
+
 		$monograph = $this->getMonograph();
 
 		// Remove the chapter author id.
-		$chapterAuthorDao = DAORegistry::getDAO('ChapterAuthorDAO');
+		$chapterAuthorDao = DAORegistry::getDAO('ChapterAuthorDAO'); /* @var $chapterAuthorDao ChapterAuthorDAO */
 		$chapterAuthorDao->deleteChapterAuthorById($author->getId(), $chapterId);
 
 		// Add it again with the correct sequence value.
 		// FIXME: primary authors not set for chapter authors.
-		$chapterAuthorDao->insertChapterAuthor($author->getId(), $chapterId, $monograph->getId(), false, $newSequence);
+		$chapterAuthorDao->insertChapterAuthor($author->getId(), $chapterId, false, $newSequence);
 	}
 
 
@@ -267,14 +314,104 @@ class ChapterGridHandler extends CategoryGridHandler {
 	// Public Chapter Grid Actions
 	//
 	/**
+	 * Edit chapter pub ids
+	 * @param $args array
+	 * @param $request PKPRequest
+	 * @return JSONMessage JSON object
+	 */
+	function identifiers($args, $request) {
+		$chapter = $this->_getChapterFromRequest($request);
+
+		import('controllers.tab.pubIds.form.PublicIdentifiersForm');
+		$form = new PublicIdentifiersForm($chapter);
+		$form->initData();
+		return new JSONMessage(true, $form->fetch($request));
+	}
+
+	/**
+	 * Update chapter pub ids
+	 * @param $args array
+	 * @param $request PKPRequest
+	 * @return JSONMessage JSON object
+	 */
+	function updateIdentifiers($args, $request) {
+		if (!$this->canAdminister($request->getUser())) return new JSONMessage(false);
+
+		$chapter = $this->_getChapterFromRequest($request);
+
+		import('controllers.tab.pubIds.form.PublicIdentifiersForm');
+		$form = new PublicIdentifiersForm($chapter);
+		$form->readInputData();
+		if ($form->validate()) {
+			$form->execute();
+			return DAO::getDataChangedEvent();
+		} else {
+			return new JSONMessage(true, $form->fetch($request));
+		}
+	}
+
+	/**
+	 * Clear chapter pub id
+	 * @param $args array
+	 * @param $request PKPRequest
+	 * @return JSONMessage JSON object
+	 */
+	function clearPubId($args, $request) {
+		if (!$request->checkCSRF()) return new JSONMessage(false);
+		if (!$this->canAdminister($request->getUser())) return new JSONMessage(false);
+
+		$chapter = $this->_getChapterFromRequest($request);
+
+		import('controllers.tab.pubIds.form.PublicIdentifiersForm');
+		$form = new PublicIdentifiersForm($chapter);
+		$form->clearPubId($request->getUserVar('pubIdPlugIn'));
+		return new JSONMessage(true);
+	}
+
+	/**
 	 * Add a chapter.
 	 * @param $args array
 	 * @param $request Request
 	 */
 	function addChapter($args, $request) {
-		// Calling editChapter() with an empty row id will add
+		if (!$this->canAdminister($request->getUser())) return new JSONMessage(false);
+		// Calling editChapterTab() with an empty row id will add
 		// a new chapter.
-		return $this->editChapter($args, $request);
+		return $this->editChapterTab($args, $request);
+	}
+
+	/**
+	 * Edit a chapter metadata modal
+	 * @param $args array
+	 * @param $request PKPRequest
+	 * @return JSONMessage JSON object
+	 */
+	function editChapter($args, $request) {
+		if (!$this->canAdminister($request->getUser())) return new JSONMessage(false);
+		$chapter = $this->_getChapterFromRequest($request);
+
+		// Check if this is a remote galley
+		$templateMgr = TemplateManager::getManager($request);
+		$templateMgr->assign(array(
+			'submissionId' => $this->getMonograph()->getId(),
+			'publicationId' => $this->getPublication()->getId(),
+			'chapterId' => $chapter->getId(),
+		));
+
+		if (array_intersect(array(ROLE_ID_MANAGER, ROLE_ID_SUB_EDITOR, ROLE_ID_ASSISTANT), $this->getAuthorizedContextObject(ASSOC_TYPE_USER_ROLES))) {
+			$publisherIdEnabled = in_array('chapter', (array) $request->getContext()->getData('enablePublisherId'));
+			$pubIdPlugins = PluginRegistry::getPlugins('pubIds');
+			$pubIdEnabled = false;
+			foreach ($pubIdPlugins as $pubIdPlugin) {
+				if ($pubIdPlugin->isObjectTypeEnabled('Chapter', $request->getContext()->getId())) {
+					$pubIdEnabled = true;
+					break;
+				}
+			}
+			$templateMgr->assign('showIdentifierTab', $publisherIdEnabled || $pubIdEnabled);
+		}
+
+		return new JSONMessage(true, $templateMgr->fetch('controllers/grid/users/chapter/editChapter.tpl'));
 	}
 
 	/**
@@ -283,12 +420,13 @@ class ChapterGridHandler extends CategoryGridHandler {
 	 * @param $request PKPRequest
 	 * @return JSONMessage JSON object
 	 */
-	function editChapter($args, $request) {
+	function editChapterTab($args, $request) {
+		if (!$this->canAdminister($request->getUser())) return new JSONMessage(false);
 		$chapter = $this->_getChapterFromRequest($request);
 
 		// Form handling
 		import('controllers.grid.users.chapter.form.ChapterForm');
-		$chapterForm = new ChapterForm($this->getMonograph(), $chapter);
+		$chapterForm = new ChapterForm($this->getMonograph(), $this->getPublication(), $chapter);
 		$chapterForm->initData();
 
 		return new JSONMessage(true, $chapterForm->fetch($request));
@@ -301,21 +439,21 @@ class ChapterGridHandler extends CategoryGridHandler {
 	 * @return JSONMessage JSON object
 	 */
 	function updateChapter($args, $request) {
+		if (!$this->canAdminister($request->getUser())) return new JSONMessage(false);
 		// Identify the chapter to be updated
 		$chapter = $this->_getChapterFromRequest($request);
 
 		// Form initialization
 		import('controllers.grid.users.chapter.form.ChapterForm');
-		$chapterForm = new ChapterForm($this->getMonograph(), $chapter);
+		$chapterForm = new ChapterForm($this->getMonograph(), $this->getPublication(), $chapter);
 		$chapterForm->readInputData();
 
 		// Form validation
 		if ($chapterForm->validate()) {
-			$chapterForm->execute($request);
-
-			$newChapter = $chapterForm->getChapter();
-
-			return DAO::getDataChangedEvent($newChapter->getId());
+			$notificationMgr = new NotificationManager();
+			$notificationMgr->createTrivialNotification($request->getUser()->getId());
+			$chapterForm->execute();
+			return DAO::getDataChangedEvent($chapterForm->getChapter()->getId());
 		} else {
 			// Return an error
 			return new JSONMessage(false);
@@ -329,19 +467,20 @@ class ChapterGridHandler extends CategoryGridHandler {
 	 * @return JSONMessage JSON object
 	 */
 	function deleteChapter($args, $request) {
+		if (!$this->canAdminister($request->getUser())) return new JSONMessage(false);
 		// Identify the chapter to be deleted
 		$chapter = $this->_getChapterFromRequest($request);
 		$chapterId = $chapter->getId();
 
 		// remove Authors assigned to this chapter first
-		$chapterAuthorDao = DAORegistry::getDAO('ChapterAuthorDAO');
+		$chapterAuthorDao = DAORegistry::getDAO('ChapterAuthorDAO'); /* @var $chapterAuthorDao ChapterAuthorDAO */
 		$assignedAuthorIds = $chapterAuthorDao->getAuthorIdsByChapterId($chapterId);
 
 		foreach ($assignedAuthorIds as $authorId) {
 			$chapterAuthorDao->deleteChapterAuthorById($authorId, $chapterId);
 		}
 
-		$chapterDao = DAORegistry::getDAO('ChapterDAO');
+		$chapterDao = DAORegistry::getDAO('ChapterDAO'); /* @var $chapterDao ChapterDAO */
 		$chapterDao->deleteById($chapterId);
 		return DAO::getDataChangedEvent();
 	}
@@ -350,10 +489,11 @@ class ChapterGridHandler extends CategoryGridHandler {
 	 * Fetch and validate the chapter from the request arguments
 	 */
 	function _getChapterFromRequest($request) {
-		$monograph = $this->getMonograph();
-		$chapterDao = DAORegistry::getDAO('ChapterDAO');
-		return $chapterDao->getChapter((int) $request->getUserVar('chapterId'), $monograph->getId());
+		return DAORegistry::getDAO('ChapterDAO')->getChapter(
+				(int) $request->getUserVar('chapterId'),
+				$this->getPublication()->getId()
+			);
 	}
 }
 
-?>
+
